requirements.txt
anthropic>=0.3.0
python-dotenv>=1.0.0
ipykernel>=6.0.0
jupyter>=1.0.0
pytest>=7.0.0
openai>=1.0.0
